internal: change variable_info.data_size type to size_t
Since this field is wrote to using sizeof and read only to be passed
down to memcpy, this change requires no other adjustment.

+++ libnopegl/src/internal.h
@@ -225,7 +225,7 @@ struct variable_opts {
 
 struct variable_info {
     void *data;
-    int data_size;
+    size_t data_size;
     int data_type;          // any of NGLI_TYPE_*
     int dynamic;
 };
